PyTorch/03-Cifar10/requirements.txt
# CIFAR-10画像分類 - 依存ライブラリ
# Python 3.7以上を推奨

torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
Pillow>=8.0.0
tensorboardX>=2.5.0
matplotlib>=3.4.0
